moved getOwner to utils folder
