+++ app/plugins/adp-backend/src/utils.ts
@@ -1,34 +1,42 @@
 import { ArmsLengthBody, DeliveryProgramme } from '@internal/plugin-adp-common';
 import { IdentityApi } from '@backstage/plugin-auth-node';
 import express from 'express';
+import { AlbRouterOptions } from './service/armsLengthBodyRouter';
 
-export function createName (name: string) {
-    const nameValue = name.replace(/\s+/g, '-').toLowerCase().substring(0,64);
-    return nameValue;
+export function createName(name: string) {
+  const nameValue = name.replace(/\s+/g, '-').toLowerCase().substring(0, 64);
+  return nameValue;
 }
 
-export function createTransformerTitle(title:string, alias?: string) {
-    const titleValue =  alias ? (title + " " + `(${alias})`) : title
-    return titleValue;
+export function createTransformerTitle(title: string, alias?: string) {
+  const titleValue = alias ? title + ' ' + `(${alias})` : title;
+  return titleValue;
 }
 
 export async function checkForDuplicateTitle(
-    store: DeliveryProgramme[] | ArmsLengthBody[],
-    title: string,
-  ): Promise<boolean> {
-    title = title.trim().toLowerCase();
-  
-    const duplicate = store.find(
-      object => object.title.trim().toLowerCase() === title,
-    );
-  
-    return duplicate !== undefined;
-  }
-  
-  export async function getCurrentUsername(
-    identity: IdentityApi,
-    req: express.Request,
-  ): Promise<string> {
-    const user = await identity.getIdentity({ request: req });
-    return user?.identity.userEntityRef ?? 'unknown';
-  }
+  store: DeliveryProgramme[] | ArmsLengthBody[],
+  title: string,
+): Promise<boolean> {
+  title = title.trim().toLowerCase();
+
+  const duplicate = store.find(
+    object => object.title.trim().toLowerCase() === title,
+  );
+
+  return duplicate !== undefined;
+}
+
+export async function getCurrentUsername(
+  identity: IdentityApi,
+  req: express.Request,
+): Promise<string> {
+  const user = await identity.getIdentity({ request: req });
+  return user?.identity.userEntityRef ?? 'unknown';
+}
+
+export function getOwner(options: AlbRouterOptions): string {
+  const { config } = options;
+  const ownerGroup = config.getConfig('rbac');
+  const owner = ownerGroup.getString('programmeAdminGroup');
+  return owner;
+}
